+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 3.3.0 (2020-04-27)
 
-- Added `NavigationExtras.state` support to the flyout component's permalink button. [#78](https://github.com/blackbaud/skyux-flyout/pull/78)
+- Added `NavigationExtras.state` support to the flyout component's permalink button. [#78](https://github.com/blackbaud/skyux-flyout/pull/78) (Thanks [@Alex-Vaky](https://github.com/Alex-Vaky)!)
 
 # 3.2.4 (2020-04-03)