+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
 ### Breaking changes
 
-- Dropped support for `rxjs@5`. Consumers may install `rxjs-compat@^6` to support older versions of `rxjs`. [#73](https://github.com/blackbaud/skyux-flyout/pull/73)
+- Dropped support for `rxjs@5`. Consumers can install `rxjs-compat@^6` to support older versions of `rxjs`. [#73](https://github.com/blackbaud/skyux-flyout/pull/73)
 
 # 3.2.4 (2020-04-03)